+++ users/src/routes/status.ts
@@ -1,8 +1,6 @@
 import express, { Request, Response, NextFunction } from 'express';
 import { body } from 'express-validator';
-import { Member } from '../models/member.model';
-import { User } from '../models/user.model';
-import { Team } from '../models/team.model';
+import { User, Team, Member } from '../models';
 import { requireUser } from '../middlewares/requireUser';
 import { validateRequest, BadRequestError } from '@cuconnex/common';
