+++ users/src/routes/team-routes/get-member.ts
@@ -1,10 +1,8 @@
 import express, { Request, Response, NextFunction } from 'express';
-import { Member } from '../../models/member.model';
-import { Team } from '../../models/team.model';
+import { Team, Member } from '../../models';
 
 import { BadRequestError, validateRequest } from '@cuconnex/common';
 import { body } from 'express-validator';
-import { requireUser } from '../../middlewares/requireUser';
 
 const router = express.Router();
